// src/tcgdb/modules/tcgdb/pokemon/price/change/yearly/dto/tcgdb.pokemon.price.change.yearly.dto.ts
export class TCGdbPokemonPricesChangeYearlyDTO {
    tcgdbPokemonPricesChangeYearly: TCGdbPokemonPriceChangeYearlyDTO[];
}

export class TCGdbPokemonPriceChangeYearlyDTO {
    tcgdbPokemonPriceChangeYearlyId: string;
    tcgdbPokemonCardId: string;
    tcgdbPokemonPriceChangeYearlyTCGPlayerId: number;
    tcgdbPokemonPriceChangeYearlySetCode: string;
    tcgdbPokemonPriceChangeYearlyCurrentLowPrice: number;
    tcgdbPokemonPriceChangeYearlyPreviousLowPrice: number;
    tcgdbPokemonPriceChangeYearlyLowPriceDifference: number;
    tcgdbPokemonPriceChangeYearlyLowPricePercentage: number;
    tcgdbPokemonPriceChangeYearlyCurrentMidPrice: number;
    tcgdbPokemonPriceChangeYearlyPreviousMidPrice: number;
    tcgdbPokemonPriceChangeYearlyMidPriceDifference: number;
    tcgdbPokemonPriceChangeYearlyMidPricePercentage: number;
    tcgdbPokemonPriceChangeYearlyCurrentHighPrice: number;
    tcgdbPokemonPriceChangeYearlyPreviousHighPrice: number;
    tcgdbPokemonPriceChangeYearlyHighPriceDifference: number;
    tcgdbPokemonPriceChangeYearlyHighPricePercentage: number;
    tcgdbPokemonPriceChangeYearlyCurrentMarketPrice: number;
    tcgdbPokemonPriceChangeYearlyPreviousMarketPrice: number;
    tcgdbPokemonPriceChangeYearlyMarketPriceDifference: number;
    tcgdbPokemonPriceChangeYearlyMarketPricePercentage: number;
    tcgdbPokemonPriceChangeYearlySubTypeName: string;
}

export class CreateTCGdbPokemonPriceChangeYearlyDTO {
    tcgdbPokemonCardId: string;
    tcgdbPokemonPriceChangeYearlyTCGPlayerId: number;
    tcgdbPokemonPriceChangeYearlySetCode: string;
    tcgdbPokemonPriceChangeYearlyCurrentLowPrice: number;
    tcgdbPokemonPriceChangeYearlyPreviousLowPrice: number;
    tcgdbPokemonPriceChangeYearlyLowPriceDifference: number;
    tcgdbPokemonPriceChangeYearlyLowPricePercentage: number;
    tcgdbPokemonPriceChangeYearlyCurrentMidPrice: number;
    tcgdbPokemonPriceChangeYearlyPreviousMidPrice: number;
    tcgdbPokemonPriceChangeYearlyMidPriceDifference: number;
    tcgdbPokemonPriceChangeYearlyMidPricePercentage: number;
    tcgdbPokemonPriceChangeYearlyCurrentHighPrice: number;
    tcgdbPokemonPriceChangeYearlyPreviousHighPrice: number;
    tcgdbPokemonPriceChangeYearlyHighPriceDifference: number;
    tcgdbPokemonPriceChangeYearlyHighPricePercentage: number;
    tcgdbPokemonPriceChangeYearlyCurrentMarketPrice: number;
    tcgdbPokemonPriceChangeYearlyPreviousMarketPrice: number;
    tcgdbPokemonPriceChangeYearlyMarketPriceDifference: number;
    tcgdbPokemonPriceChangeYearlyMarketPricePercentage: number;
    tcgdbPokemonPriceChangeYearlySubTypeName: string;
}